+++ packages/compiler/src/init/init.ts
@@ -4,6 +4,7 @@ import pc from "picocolors";
 import prompts from "prompts";
 import * as semver from "semver";
 import { stringify } from "yaml";
+import { builtInTemplates } from "../../templates/index.js";
 import { TypeSpecConfigFilename } from "../config/config-loader.js";
 import { formatTypeSpec } from "../core/formatter.js";
 import { createDiagnostic } from "../core/messages.js";
@@ -19,7 +20,6 @@ import {
   InitTemplateLibrarySpec,
   InitTemplateSchema,
 } from "./init-template.js";
-import { builtInTemplates } from "./templates/index.js";
 
 export interface ScaffoldingConfig extends InitTemplate {
   /**
